Fix flaky test.
An assertion was removed that assumed that in async training, the global step watcher will not have started before the workers have passed their execution barrier. This assumption is not correct, as its possible (but rare) for each worker to have finished the warmup, then have the global step watcher start, before the execution barrier is reached.

PiperOrigin-RevId: 195145396