If OS ENV are overridden, a warning msg is printed
Updated params.num_intra_threads condition. This is because if for example OS ENV OMP_NUM_THREADS is set 15, and --num_intra_threads is not passed, then TF is picking up OS ENV and sets the value to 15. This update will set the OS ENV OMP_NUM_THREADS to the defaults and if --num_intra_threads is not passed or if value < 1 is passed.