Give better error message with OutOfRangeError.
Before, if an OutOfRangeError was thrown, the Supervisor would silently ignore it, causing the cryptic error: "UnboundLocalError: local variable 'num_steps' referenced before assignment". See issues #150 and #197. This change adds a try-catch to catch the OutOfRangeError, wrapping it with a RuntimeError, so that Supervisor does not ignore it.

Also, everything under the managed_session context manager has been moved to a new function, to avoid deep nesting.

PiperOrigin-RevId: 217622216