Updating `get_gpu_info()` to support AMD GPUs
This PR/commit updates the `get_gpu_info` routine (within `utils.py`) to support AMD GPUs.

The `get_gpu_info` routine now checks for the present of `rocm-smi` (equivalent of `nvidia-smi` for AMD GPUs) in the PATH
If found, then it will call `rocm-smi` to get the AMD GPU specific info,
Else if will call the `nvidia-smi` to get the NVIDIA GPU specific info.

Similar to the case for NVIDIA GPUs, the code assumes that if the system has multiple GPUs, then they are all the same.